perf: reduce chat input lag via React.memo and polling hook extraction
- Wrap MessageComposer with React.memo to prevent re-renders caused by
  parent state updates (agent statuses, context usage polling)
- Extract fetchStatuses into useAgentStatuses hook (use-agent-statuses.ts)
- Extract fetchContextUsage into useContextUsage hook (use-context-usage.ts)
- Remove both polling setIntervals from _layout.chat.tsx so status/context
  updates no longer propagate re-renders to MessageComposer
